Create autohibernate.sh
This script is intended to run on AWS Linux within an EC2 instance. Otherwise it will not be able to call its instance-id from the IP provided.
This script is intended to hibernate the instance its running on if no user has been present for 30 minutes after the last check.